Commented out code for book Table of Contents
Instead of a book version of a Table of Contents (which in hindsight, probably wasn't the best of ideas), the table of contents will instead be on the scalar index page. All public articles will be shown. They will be in chronological order, with a description, and showing the author's name (just like in the original version of Scalar).